# swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "나만의 BE 커뮤니티",
    "description": "Node.js Community",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/register": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "nickname": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "confirmPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "작성됨"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "412": {
            "description": "사전조건 실패"
          }
        }
      }
    },
    "/api/auth": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          }
        }
      }
    },
    "/api/posts": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "권한 없음"
          },
          "412": {
            "description": "사전조건 실패"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          }
        }
      }
    },
    "/api/posts/{post_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "권한 없음"
          },
          "412": {
            "description": "사전조건 실패"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "권한 없음"
          },
          "404": {
            "description": "찾을 수 없음"
          }
        }
      }
    },
    "/api/comments/{post_id}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "권한 없음"
          },
          "412": {
            "description": "사전조건 실패"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/api/comments/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "권한 없음"
          },
          "412": {
            "description": "사전조건 실패"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "권한 없음"
          },
          "404": {
            "description": "찾을 수 없음"
          }
        }
      }
    },
    "/api/posts/{post_id}/like": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "권한 없음"
          },
          "404": {
            "description": "찾을 수 없음"
          }
        }
      }
    },
    "/api/likes/posts": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "권한 없음"
          }
        }
      }
    }
  }
}